make info field a trait
