+++ src/info/head.rs
