+++ src/info/info_field.rs
@@ -1,6 +1,28 @@
+pub struct InfoFieldValue {
+    pub r#type: InfoFieldType,
+    pub value: String,
+}
+
+pub trait InfoField {
+    fn value(&self) -> InfoFieldValue;
+}
+
+pub trait InfoFieldValueGetter {
+    fn get(&self, disabled_fields: &[InfoFieldType]) -> Option<InfoFieldValue>;
+}
+
+impl InfoFieldValueGetter for dyn InfoField {
+    fn get(&self, disabled_fields: &[InfoFieldType]) -> Option<InfoFieldValue> {
+        let infor_field = self.value();
+        if disabled_fields.contains(&infor_field.r#type) || infor_field.value.is_empty() {
+            return None;
+        }
+        Some(infor_field)
+    }
+}
+
 #[derive(Clone, clap::ValueEnum, Debug, Eq, PartialEq)]
-pub enum InfoField {
-    Title,
+pub enum InfoFieldType {
     Project,
     Head,
     Pending,
@@ -18,82 +40,24 @@ pub enum InfoField {
     License,
 }
 
-#[derive(Default)]
-pub struct InfoFieldOff {
-    pub title: bool,
-    pub project: bool,
-    pub head: bool,
-    pub pending: bool,
-    pub version: bool,
-    pub created: bool,
-    pub languages: bool,
-    pub dependencies: bool,
-    pub authors: bool,
-    pub last_change: bool,
-    pub contributors: bool,
-    pub repo: bool,
-    pub commits: bool,
-    pub lines_of_code: bool,
-    pub size: bool,
-    pub license: bool,
-}
-
-impl From<&Vec<InfoField>> for InfoFieldOff {
-    fn from(fields_to_hide: &Vec<InfoField>) -> Self {
-        let mut info_field_off = InfoFieldOff {
-            ..Default::default()
-        };
-
-        for field in fields_to_hide.iter() {
-            match field {
-                InfoField::Title => info_field_off.title = true,
-                InfoField::Project => info_field_off.project = true,
-                InfoField::Head => info_field_off.head = true,
-                InfoField::Pending => info_field_off.pending = true,
-                InfoField::Version => info_field_off.version = true,
-                InfoField::Created => info_field_off.created = true,
-                InfoField::Languages => info_field_off.languages = true,
-                InfoField::Dependencies => info_field_off.dependencies = true,
-                InfoField::Authors => info_field_off.authors = true,
-                InfoField::LastChange => info_field_off.last_change = true,
-                InfoField::Contributors => info_field_off.contributors = true,
-                InfoField::Repo => info_field_off.repo = true,
-                InfoField::Commits => info_field_off.commits = true,
-                InfoField::LinesOfCode => info_field_off.lines_of_code = true,
-                InfoField::Size => info_field_off.size = true,
-                InfoField::License => info_field_off.license = true,
-            }
+impl InfoFieldType {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            InfoFieldType::Project => "Project",
+            InfoFieldType::Head => "HEAD",
+            InfoFieldType::Pending => "Pending",
+            InfoFieldType::Version => "Version",
+            InfoFieldType::Created => "Create",
+            InfoFieldType::Languages => "Languages",
+            InfoFieldType::Dependencies => "Dependencies",
+            InfoFieldType::Authors => "Authors",
+            InfoFieldType::LastChange => "Last Change",
+            InfoFieldType::Contributors => "Contributors",
+            InfoFieldType::Repo => "Repo",
+            InfoFieldType::Commits => "CommitS",
+            InfoFieldType::LinesOfCode => "Lines of code",
+            InfoFieldType::Size => "Size",
+            InfoFieldType::License => "License",
         }
-
-        info_field_off
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_from_vec_of_info_fields() {
-        let fields_to_hide = vec![InfoField::Version, InfoField::Repo];
-        let info_field_off = InfoFieldOff::from(&fields_to_hide);
-
-        assert!(!info_field_off.title);
-        assert!(!info_field_off.project);
-        assert!(!info_field_off.head);
-        assert!(!info_field_off.pending);
-        assert!(!info_field_off.pending);
-        assert!(info_field_off.version);
-        assert!(!info_field_off.created);
-        assert!(!info_field_off.languages);
-        assert!(!info_field_off.dependencies);
-        assert!(!info_field_off.authors);
-        assert!(!info_field_off.last_change);
-        assert!(!info_field_off.contributors);
-        assert!(info_field_off.repo);
-        assert!(!info_field_off.commits);
-        assert!(!info_field_off.lines_of_code);
-        assert!(!info_field_off.size);
-        assert!(!info_field_off.license);
     }
 }
